Fix Cipher#pkcs5_keyivgen by passing ctx cipher
pkcs5_keyivgen wrongly passes self's ctx as the first parameter of
EVP_BytesToKey, causing it to fail. The first parameter needs to be a
pointer to EVP_MD, which is a libcrypto cipher description.

This additionally changes salt from a default null pointer, which would
attempt to call the private bytesize method.